+++ awscli_saml_sso/__init__.py
@@ -1,5 +1,3 @@
 import logging
 
-logging.basicConfig(level=logging.INFO)
-
 __version__ = "0.2.0"