+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+HEAD (Unreleased)
+-----------------
+
+* Configure default log level to WARN in order to get printed text in terminal uncluttered by logs
+* Add a ``--log-level`` option and corresponding ``ASS_LOG_LEVEL`` environment variable to override default log level
+
 0.2.0 (2020-12-04)
 ------------------